requirements.txt
beautifulsoup4
diskcache
python-dotenv
playwright
#selenium
#webdriver-manager
